Improve float parser
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-libs-2.1.1@196 75748be7-0387-4326-94c7-f41ce0883d2c

+++ utils/iniParser.h
@@ -30,7 +30,7 @@ class iniParser {
 public:
 	static long parseLong(const char* str);
 
-	static float parseFloat(const char* str);
+	static double parseDouble(const char* str);
 
 private:
 	typedef std::map<std::string, std::string> keys_t;
